+++ src/main/java/com/gregtechceu/gtceu/data/recipe/CraftingComponent.java
@@ -12,8 +12,8 @@ import com.gregtechceu.gtceu.common.data.GTItems;
 import com.gregtechceu.gtceu.common.data.GTMachines;
 import com.gregtechceu.gtceu.common.data.GTMaterials;
 import com.gregtechceu.gtceu.data.recipe.event.CraftingComponentModificationEvent;
-import com.gregtechceu.gtceu.integration.kjs.GTCEuServerEvents;
-import com.gregtechceu.gtceu.integration.kjs.events.CraftingComponentEventJS;
+import com.gregtechceu.gtceu.integration.kjs.GTCEuStartupEvents;
+import com.gregtechceu.gtceu.integration.kjs.events.CraftingComponentsEventJS;
 
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Blocks;
@@ -965,7 +965,7 @@ public class CraftingComponent {
     private static final class KJSCallWrapper {
 
         private static void craftingComponentModification() {
-            GTCEuServerEvents.CRAFTING_COMPONENTS.post(new CraftingComponentEventJS());
+            GTCEuStartupEvents.CRAFTING_COMPONENTS.post(new CraftingComponentsEventJS());
         }
     }
 }
